[Fix] must save meta file after checkpoint_list remove

+++ xtuner/v1/train/trainer.py
@@ -798,13 +798,17 @@ class Trainer:
                     )
                 )
 
+        if self._checkpoint_maxkeep > 0 and len(current_exp.checkpoint_list) > self._checkpoint_maxkeep:
+            deleted_checkpoints = current_exp.checkpoint_list[: -self._checkpoint_maxkeep]
+            current_exp.checkpoint_list = current_exp.checkpoint_list[-self._checkpoint_maxkeep :]
+            for ckp_dir in deleted_checkpoints:
+                if self.rank == 0 and Path(ckp_dir).exists():
+                    rmtree(ckp_dir)
+        
+        # Must save meta after deleting checkpoints to ensure the checkpoint_list is updated in the meta file
+        if self.rank == 0:
             with meta_path.open("w") as f:
                 f.write(self.meta.model_dump_json(indent=2))
-
-        if self._checkpoint_maxkeep > 0 and len(current_exp.checkpoint_list) > self._checkpoint_maxkeep:
-            ckpt_to_remove = current_exp.checkpoint_list.pop(0)
-            if self.rank == 0:
-                rmtree(ckpt_to_remove)
 
         dist.barrier()
 
@@ -1070,7 +1074,7 @@ class Trainer:
             deleted_hf_checkpoints = self.meta.latest_exp.hf_checkpoint_list[: -self._hf_max_keep]
             self.meta.latest_exp.hf_checkpoint_list = self.meta.latest_exp.hf_checkpoint_list[-self._hf_max_keep :]
             for hf_dir in deleted_hf_checkpoints:
-                if self.rank == 0:
+                if self.rank == 0 and Path(hf_dir).exists():
                     rmtree(hf_dir)
 
         self._engine.save_hf(str(save_hf_path))
